Refactored Reply feature to separate component
