+++ src/main/css/components/timeline/TweetBlock.scss
@@ -34,20 +34,20 @@ $names-padding: ($user-div-width - $names-width) / 2;
 		bottom: 0;
 		z-index: 7;
 
-		.reply-modal {
+		.modal-wrapper {
 			box-sizing: border-box;
 			border: 1px solid black;
-			width: 500px;
+			width: 50%;
 			z-index: 10;
 			position: fixed;
 			right: 0;
 			top: 0;
 			bottom: 0;
 			margin: 10px;
-			padding: 10px;
+			padding: 20px;
 			background-color: white;
 
-			.modal-body {
+			.modal-content {
 				display: flex;
 				flex-direction: column;
 				height: 100%;
@@ -96,7 +96,7 @@ $names-padding: ($user-div-width - $names-width) / 2;
 			.reply-modal-buttons-div {
 				display: flex;
 				justify-content: flex-end;
-				margin-top: 5px;
+				margin-top: 20px;
 				.close-reply-modal-button {
 					margin-right: 10px;
 				}
